% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_rcppeigen_internal}
\alias{square_rcppeigen_internal}
\title{Calculate squares.}
\usage{
square_rcppeigen_internal(x)
}
\arguments{
\item{x}{A vector of numbers to square.}
}
\value{
Square of x.
}
\description{
Calculate squares.
}
